requirements.txt
yfinance==0.1.63
plotly==5.1.0
numpy==1.19.5
pandas==1.2.4
streamlit==0.84.0
Pillow==8.3.1
